+++ kite/src/main/kotlin/com/tang/kite/session/SqlSession.kt
@@ -27,8 +27,6 @@ interface SqlSession : AutoCloseable {
 
     fun <T> batchInsert(method: Method, mapperInterface: Class<T>, parameter: Any, batchSize: Int): Int
 
-    fun <T> batchInsertSelective(method: Method, mapperInterface: Class<T>, parameter: Any, batchSize: Int): Int
-
     fun <T> update(method: Method, mapperInterface: Class<T>, parameter: Any): Int
 
     fun <T> update(method: Method, mapperInterface: Class<T>, parameter: Any, condition: Any): Int
@@ -40,8 +38,6 @@ interface SqlSession : AutoCloseable {
     fun <T> updateWrapper(method: Method, mapperInterface: Class<T>, type: Class<T>, parameter: Any): Int
 
     fun <T> batchUpdate(method: Method, mapperInterface: Class<T>, parameter: Any, batchSize: Int): Int
-
-    fun <T> batchUpdateSelective(method: Method, mapperInterface: Class<T>, parameter: Any, batchSize: Int): Int
 
     fun <T> delete(method: Method, mapperInterface: Class<T>, type: Class<T>, parameter: Any): Int